fix(review): offline-first suppression, durable retries, late checkpoint, thinner agent
Addresses the Greptile review on #27:
- The offline check now outranks critical priority, so a low-battery
  evaluation with no connection suppresses instead of burning its
  30-minute cooldown on a card nobody hears; the next telemetry after
  reconnect retries promptly.
- A deferred utterance whose retry lands in a suppressing moment
  re-schedules itself every 30 minutes, bounded at 5 attempts — a
  changelog deferred overnight no longer vanishes if the device is
  offline at 09:00.
- The consolidation checkpoint (lastProcessedLeafId) is written only
  after the block rewrite, prompt refresh, and memory inserts succeed;
  a failure mid-run reprocesses the same window, which the content
  dedupe makes idempotent.
- The firmware lifecycle and the nightly consolidation move out of the
  Apollo agent into src/ota/lifecycle.ts and src/memory/nightly.ts with
  injected dependencies; the agent keeps thin guards and wiring.
